[simplify] move {consul,docker}2->{consul,docker}
